api/requirements.txt
flask-restful==0.3.8
flask==1.1.2
grpcio==1.35.0
grpcio-tools==1.35.0
pydantic==1.7.3
email-validator==1.1.2
gunicorn==20.0.4